Update ADI to use actual Frobenius norm
The spectral norms of `L*D*L'` and `L'L*D` coincide. Due to the
equivalence of all operator norms, one may simply replace the spectral
norm by the Frobenius norm, which is much cheaper to compute. However,
this reasoning is somewhat hand-wavy, as the scaling factor includes the
dimension of `D`. Therefore, update the ADI to use the actual Frobenius
norm of the low-rank `L*D*L'`, which should be no more expensive to
compute than the quantity it replaces.

The motivation behind this change is to make the ratio

    norm(initial_residual) / norm(prob.Q)

easier to interpret, as it may now be understood as a normalized
residual. Note that `initial_residual` and `prob.Q` may not have the
same "inner dimension".