Add naive Armijo line search
... and add `observe_gare_metadata!`.

The way the Riccati residual is computed is very naive. A more robust
implementation should instead adapt formula (5.2b) from Benner et al.
for the LRSIF. However, as the line search is mostly performed for early
Newton iterations, all the matrices involved have relatively few
columns. Therefore, a naive implementation doesn't really hurt here.
From some numerical experiments with Rail(1357) and a heavy weighting of
the quadratic part; `G = LDLᵀ(1000*B, I)`; the line search takes less
than 50ms while all the other parts of the ADI tally to almost 10s,
which is roughly 0.5% of the runtime.

Adapting the exact line search for the LRSIF is a bit more involved.
More specifically, the formulas given in section 5.3 of Benner et al.
are not immediately applicable if the Lyapunov residual is in LRSIF.