Add GPU support for Ros1
* Add `orthf()`: orthogonal-times-square factorization
* Update `Stuff.orth()` to use SVD instead of QR
* Remove `_factorize`; performance for sparse D in LDL^T should be worse now

The perhaps biggest hurdle was to find a suitable GPU algorithm to
obtain a orthogonal-times-square factorization, which is needed for the
column compression and norm computation of an LDL^T factorization.
There is no out-of-the-box QR in CUDA.jl for sparse matrices. Instead of
implementing my own, I use the more expensive SVD.

The user can control where certain parts of the algorithm are executed
via the type of the system matrices (E, A, B, C) as well as the initial
solution X0 = LDL^T.

* E, L, and D on the XPU: everything runs on the XPU
* E, L on the GPU, D on the CPU: everything runs on the GPU,
  except for sub-problems projected onto the space of D.
  Namely: eigen-decomposition and parts of the LDL^T norm are
  computed on the CPU.